final commit

+++ SBS19010/src/main/java/ie/cct/SBS19010/AverageWeightResponse.java
@@ -1,30 +1,32 @@
 package ie.cct.SBS19010;
 
 public class AverageWeightResponse {
-	
+
 	String type;
-	Float avgWeight;
-	
-	AverageWeightResponse(){}
-	
-	
-	public AverageWeightResponse(String type, Float avgWeight) {
+	Double avgWeight;
+
+	AverageWeightResponse() {
+	}
+
+	public AverageWeightResponse(String type, Double avgWeight) {
 		super();
 		this.type = type;
 		this.avgWeight = avgWeight;
 	}
 
-
 	public String getType() {
 		return type;
 	}
+
 	public void setType(String type) {
 		this.type = type;
 	}
-	public Float getAvgWeight() {
+
+	public Double getAvgWeight() {
 		return avgWeight;
 	}
-	public void setAvgWeight(Float avgWeight) {
+
+	public void setAvgWeight(Double avgWeight) {
 		this.avgWeight = avgWeight;
 	}
